+++ modules/BiulabWidget/MapEditorBase.hh
@@ -1,0 +1,14 @@
+#pragma once
+#ifndef MAP_EDITOR_BASE_H_
+#define MAP_EDITOR_BASE_H_
+
+#include "CoreBase.hh"
+
+class MapEditorBase {
+  public:
+    MapEditorBase() {}
+
+    ~MapEditorBase() {}
+};
+
+#endif