+++ modules/MapEditor/MapEditor.cc
@@ -18,11 +18,12 @@ MapEditor::MapEditor()
     , tools(T_NONE)
     , selections(S_NONE)
     , selectPosCnt(0) {
-    infos.insert({"T_NONE", {"None", {0.0, 0.0, 0.0, 1.0}}});
-    infos.insert({"T_EDGE", {"Edge", {1.0, 1.0, 1.0, 0.4}}});
-    infos.insert({"T_RUBBISH", {"Rubbish", {1.0, 0.0, 0.0, 0.5}}});
-    infos.insert({"T_EMPTY", {"Empty", {1.0, 1.0, 1.0, 0.1}}});
-    infos.insert({"T_OUT", {"Out", {1.0, 1.0, 1.0, 0.0}}});
+    infos.insert({int(T_NONE), {"None", {0.0, 0.0, 0.0, 1.0}}});
+    infos.insert({int(T_EDGE), {"Edge", {1.0, 1.0, 1.0, 0.4}}});
+    infos.insert({int(T_RUBBISH), {"Rubbish", {1.0, 0.0, 0.0, 0.5}}});
+    infos.insert({int(T_EMPTY), {"Empty", {1.0, 1.0, 1.0, 0.1}}});
+    infos.insert({int(T_OUT), {"Out", {1.0, 1.0, 1.0, 0.0}}});
+    infos.insert({int(T_GEM), {"Gem", {0.0, 1.0, 0.0, 0.5}}});
 }
 
 MapEditor::~MapEditor() {}
@@ -36,6 +37,8 @@ void MapEditor::CheckAndSetElem(SimpleMap& mp, vec_2i pos) {
         mp.SetElem(pos, MAP_RUBBISH);
     } else if (tools == T_OUT) {
         mp.SetElem(pos, MAP_OUT);
+    } else if (tools == T_GEM) {
+        mp.SetElem(pos, MAP_GEM);
     }
 }