+++ modules/MapEditor/MapEditorModule.hh
@@ -3,6 +3,7 @@
 #define MAP_EDITOR_MODULE_H_
 
 #include "DataManager.hh"
+#include "DynamicClass.hh"
 #include "InspectWindow.hh"
 #include "MapEditorWindow.hh"
 #include "ModuleBase.hh"
@@ -10,22 +11,18 @@
 #include "SimpleMapWindow.hh"
 
 typedef struct MapEditorData {
-    InspectInfo*  inspect_info;
-    OverviewInfo* overview_info;
-    MapEditor*    map_editor;
-    SimpleMap*    simple_map;
+    MapEditorWindow* map_editor_window;
+    SimpleMapWindow* simple_map_window;
+    OverviewWindow*  overview_window;
+    InspectWindow*   insepect_window;
+    InspectInfo*     inspect_info;
+    OverviewInfo*    overview_info;
+    MapEditor*       map_editor;
+    SimpleMap*       simple_map;
 } MapEditorData;
 
 class MapEditorModule : public ModuleBase {
   private:
-    MapEditorWindow map_editor_window;
-
-    SimpleMapWindow simple_map_window;
-
-    OverviewWindow overview_window;
-
-    InspectWindow insepect_window;
-
   public:
     MapEditorData* data;