+++ modules/SimpleMap/SimpleMapConfig.hh
@@ -6,5 +6,6 @@
 #define MAP_WIDTH 16   // 地图宽度
 #define MAP_HEIGHT 16  // 地图长度
 #define RUBBISH_CNT 10 // 垃圾数量
+#define GEM_CNT 0      // 宝石数量
 
 #endif